+++ src/Cart/Changelog.md
@@ -8,6 +8,7 @@
 - Dropped PHP 8.0 & PHP 8.1 Support
 - Dropped Laravel 9 Support
 - Dropped Enum v3 Support
+- Changed minimal Enum requirement to v4.1
 
 ## 3.x Series