+++ src/Cart/Changelog.md
@@ -11,6 +11,7 @@
 - Added PHP 8.3 Support
 - Changed minimum Laravel version to v10.38.2
 - Changed minimal Enum requirement to v4.2
+- Removed the throwing of `CartUpdated` event when destroying a cart (`CartDeleting` and `CartDeleted` remains)
 - BC: Changed the `CheckoutSubjectItem` interface into Configurable & Schematized
 - BC: Added argument and return types to all `Cart` and `CartManager` interface methods