+++ src/Cart/Changelog.md
@@ -9,6 +9,7 @@
 - Dropped Laravel 9 Support
 - Dropped Enum v3 Support
 - Added PHP 8.3 Support
+- Added Laravel 11 Support
 - Changed minimum Laravel version to v10.38.2
 - Changed minimal Enum requirement to v4.2
 - Removed the throwing of `CartUpdated` event when destroying a cart (`CartDeleting` and `CartDeleted` remains)