+++ src/Cart/Changelog.md
@@ -11,7 +11,7 @@
 - Added PHP 8.3 Support
 - Added Laravel 11 Support
 - Added Cart item configuration support (different configurations constitute separate cart items) to the `Cart::addItem()` method
-- Changed minimum Laravel version to v10.38.2
+- Changed minimum Laravel version to v10.43
 - Changed minimal Enum requirement to v4.2
 - Removed the throwing of `CartUpdated` event when destroying a cart (`CartDeleting` and `CartDeleted` remains)
 - BC: Changed the `CheckoutSubjectItem` interface into Configurable & Schematized